pyproject.toml
[project]
name = "gmail-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "composio-langgraph>=0.7.19",
    "langchain[anthropic]>=0.3.26",
    "langgraph>=0.5.0",
    "langsmith>=0.4.3",
    "mcp-python-client>=0.1.9",
    "python-dotenv>=1.1.1",
]
